Support to path watches and master namespace
Here are the motivations:

1) Path watches

In some projects, whole directories may contain snipets of CSS code.
In these cases, adding each namespace to the watching list can be a
hassle. Therefore, this proposal allows users to specify paths on the
watch list alongside namespaces

2) Master namespace

In some projects, a change on a file representing a CSS component
dowstream may require a "master namespace" to be recompiled. This
proposal makes this possible.

How it works:

The interface for `new-garden-watcher` is now a map that takes two
keywords a `:watch` (mandatory) and a `:master-ns`.

`:watch` is the same vector of the previous version but now it also
accepts paths. Paths are differentiated from namespaces by being
represented in strings (instead of symbols for namespaces).

The optional `:master-ns` takes a namespace name in a symbol. This
namespace will allways be recompiled shall anything change.

The interface for `garden-watcher.main` also changed. Now a flag `-n`
or `--namespace` can be used to specify namespaces. Paths are a
possible future extension.